+++ src/Orchard.Cms.Web/Modules/Orchard.Body/Views/BodyPart.Summary.cshtml
@@ -1,6 +1,6 @@
 ﻿@model Orchard.Body.ViewModels.BodyPartViewModel
 @inject Orchard.Tokens.Services.ITokenizer Tokenizer
-@using Orchard.Utility
+@using Microsoft.AspNetCore.Mvc.Modules.Utilities
 
 @{
     string body = Model.Body;
