+++ src/Orchard.Cms.Web/Modules/Orchard.ContentTypes/Controllers/AdminController.cs
@@ -16,7 +16,7 @@ using Orchard.ContentTypes.ViewModels;
 using Orchard.DisplayManagement.ModelBinding;
 using Orchard.DisplayManagement.Notify;
 using Orchard.Environment.Shell;
-using Orchard.Utility;
+using Microsoft.AspNetCore.Mvc.Modules.Utilities;
 using YesSql.Core.Services;
 using System.Reflection;
 using Microsoft.AspNetCore.Mvc.Modules.ActionConstraints;
