+++ src/Orchard.Cms.Web/Modules/Orchard.ContentTypes/ViewModels/EditPartViewModel.cs
@@ -2,7 +2,7 @@
 using Microsoft.AspNetCore.Mvc.ModelBinding;
 using Newtonsoft.Json.Linq;
 using Orchard.ContentManagement.Metadata.Models;
-using Orchard.Utility;
+using Microsoft.AspNetCore.Mvc.Modules.Utilities;
 
 namespace Orchard.ContentTypes.ViewModels
 {
