+++ src/Orchard.Cms.Web/Modules/Orchard.ContentTypes/Views/Admin/AddFieldTo.cshtml
@@ -1,4 +1,4 @@
-@using Orchard.Utility
+@using Microsoft.AspNetCore.Mvc.Modules.Utilities
 @using Orchard.ContentManagement.Metadata.Models;
 
 @model AddFieldViewModel
