+++ src/Orchard.Cms.Web/Modules/Orchard.ContentTypes/Views/Admin/EditPart.cshtml
@@ -1,6 +1,6 @@
 @model EditPartViewModel
 @using Orchard.ContentManagement.Metadata.Models
-@using Orchard.Utility
+@using Microsoft.AspNetCore.Mvc.Modules.Utilities
 
 @{
     var orderedFields = Model.PartDefinition.Fields
