+++ src/Orchard.Cms.Web/Modules/Orchard.Contents/Views/Admin/Display.cshtml
@@ -1,5 +1,5 @@
 ﻿@using Orchard.ContentManagement;
-@using Orchard.Utility;
+@using Microsoft.AspNetCore.Mvc.Modules.Utilities;
 @{
     ContentItem contentItem = Model.ContentItem;
     //Html.AddPageClassNames("detail-" + contentItem.ContentType.HtmlClassify());
