+++ src/Orchard.Cms.Web/Modules/Orchard.Contents/Views/Content.DetailAdmin.cshtml
@@ -1,4 +1,4 @@
-﻿@using Orchard.Utility;
+﻿@using Microsoft.AspNetCore.Mvc.Modules.Utilities;
 @using Orchard.ContentManagement
 @inject IContentManager ContentManager
