+++ src/Orchard.Cms.Web/Modules/Orchard.Contents/Views/Content.Summary.cshtml
@@ -1,4 +1,4 @@
-﻿@using Orchard.Utility;
+﻿@using Microsoft.AspNetCore.Mvc.Modules.Utilities;
 @{
     var contentTypeClassName = ((string)Model.ContentItem.ContentType).HtmlClassify();
 }
