+++ src/Orchard.Cms.Web/Modules/Orchard.Contents/Views/Content.cshtml
@@ -1,4 +1,4 @@
-﻿@using Orchard.Utility;
+﻿@using Microsoft.AspNetCore.Mvc.Modules.Utilities;
 
 @{
