+++ src/Orchard.Cms.Web/Modules/Orchard.Contents/Views/Item/Preview.cshtml
@@ -1,5 +1,5 @@
 ﻿@using Orchard.ContentManagement
-@using Orchard.Utility
+@using Microsoft.AspNetCore.Mvc.Modules.Utilities
 @{
     ContentItem contentItem = Model.ContentItem;
     //Html.AddPageClassNames("preview", "detail-" + contentItem.ContentType.HtmlClassify());
