+++ src/Orchard.Cms.Web/Modules/Orchard.Menu/Views/Admin/Edit.cshtml
@@ -1,6 +1,6 @@
 @using Orchard.ContentManagement
 @using Orchard.ContentManagement.MetaData
-@using Orchard.Utility
+@using Microsoft.AspNetCore.Mvc.Modules.Utilities
 
 @inject IContentDefinitionManager ContentDefinitionManager
