+++ src/Orchard.Cms.Web/Modules/Orchard.Settings/Views/Admin/Index.cshtml
@@ -1,5 +1,5 @@
 ﻿@model AdminIndexViewModel
-@using Orchard.Utility
+@using Microsoft.AspNetCore.Mvc.Modules.Utilities
 
 <h1>@RenderTitleSegments(T["Settings"])</h1>
