+++ src/Orchard.Cms.Web/Modules/Orchard.Settings/Views/SiteSettings.SaveButton.cshtml
@@ -1,4 +1,4 @@
-﻿@using Orchard.Utility
+﻿@using Microsoft.AspNetCore.Mvc.Modules.Utilities
 <button class="primaryAction btn btn-primary" type="submit">@T["Save"]</button>
 
 @{
