+++ src/Orchard.ContentManagement.Abstractions/project.json
@@ -2,7 +2,7 @@
   "version": "2.0.0-*",
   "dependencies": {
     "Microsoft.AspNetCore.Html.Abstractions": "1.1.0",
-    "Microsoft.AspNetCore.Mvc.Modules.Abstractions": "1.0.0-*",
+    "Microsoft.AspNetCore.Mvc.Modules": "1.0.0-*",
     "Microsoft.AspNetCore.Routing.Abstractions": "1.1.0",
     "Newtonsoft.Json": "9.0.1",
     "NETStandard.Library": "1.6.1"
